requirements.txt
numpy>=1.23
scipy>=1.9
pandas>=1.5
scikit-learn>=1.1
torch>=1.13
torchvision>=0.14
matplotlib>=3.6
seaborn>=0.12
networkx>=2.8
pyyaml>=6.0
skyfield>=1.46
astropy>=5.2
